add event comment #22 close #22

+++ BabyCare/app/src/main/java/com/ben/yjh/babycare/main/event/CommentActivity.java
@@ -3,24 +3,36 @@ package com.ben.yjh.babycare.main.event;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.v4.widget.NestedScrollView;
+import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
+import android.text.Editable;
+import android.text.TextWatcher;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.EditText;
+import android.widget.ImageButton;
+import android.widget.TextView;
 
 import com.ben.yjh.babycare.R;
 import com.ben.yjh.babycare.base.BaseActivity;
 import com.ben.yjh.babycare.http.EventTaskHandler;
 import com.ben.yjh.babycare.http.HttpResponseInterface;
 import com.ben.yjh.babycare.model.CommentsResult;
+import com.ben.yjh.babycare.model.Event;
 import com.ben.yjh.babycare.model.EventComment;
+import com.ben.yjh.babycare.model.EventLike;
 import com.ben.yjh.babycare.model.HttpBaseResult;
 import com.ben.yjh.babycare.util.Constants;
 import com.ben.yjh.babycare.util.Utils;
+import com.ben.yjh.babycare.widget.recyclerview.LoadMoreListener;
+import com.ben.yjh.babycare.widget.recyclerview.LoadMoreRecyclerView;
+import com.ben.yjh.babycare.widget.recyclerview.ProgressView;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -33,7 +45,10 @@ public class CommentActivity extends BaseActivity implements CommentAdapter.Comm
     private List<EventComment> mComments;
     private int mEventId;
     private EventComment mReplyComment;
-    private RecyclerView mRecyclerView;
+    private LoadMoreRecyclerView mRecyclerView;
+    private SwipeRefreshLayout mSwipeRefreshLayout;
+    private String mNextUrl;
+    private ImageButton mSendButton;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -46,24 +61,119 @@ public class CommentActivity extends BaseActivity implements CommentAdapter.Comm
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setDisplayShowHomeEnabled(true);
         setStatusBarMargin(R.id.content_layout);
+        mSendButton = (ImageButton) findViewById(R.id.ib_send);
+        mSendButton.setOnClickListener(this);
         mRootView = findViewById(R.id.content_layout);
         mRootView.setFitsSystemWindows(true);
-        mRecyclerView = (RecyclerView) findViewById(R.id.recycler_view);
+        mRecyclerView = (LoadMoreRecyclerView) findViewById(R.id.recycler_view);
         mRecyclerView.setLayoutManager(new LinearLayoutManager(this));
         mRecyclerView.setHasFixedSize(false);
         mRecyclerView.setNestedScrollingEnabled(false);
+        ProgressView progressView = new ProgressView(this);
+        progressView.setIndicatorId(ProgressView.BallPulse);
+        progressView.setIndicatorColor(0xff69b3e0);
+        mRecyclerView.setFootLoadingView(progressView);
+
+        TextView textView = new TextView(this);
+        textView.setText(R.string.load_finish);
+        mRecyclerView.setFootEndView(textView);
+
         mCommentEditText = (EditText) findViewById(R.id.et_comment);
         mCommentEditText.setHint(R.string.add_comment);
 //        mCommentEditText.requestFocus();
+        if (mCommentEditText.getText() == null ||
+                mCommentEditText.getText().toString().trim().isEmpty()) {
+            mSendButton.setEnabled(false);
+        }
+        mCommentEditText.addTextChangedListener(new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+
+            }
+
+            @Override
+            public void onTextChanged(CharSequence s, int start, int before, int count) {
+
+            }
+
+            @Override
+            public void afterTextChanged(Editable s) {
+                if (s == null || s.toString().trim().isEmpty()) {
+                    mSendButton.setEnabled(false);
+                } else {
+                    mSendButton.setEnabled(true);
+                }
+            }
+        });
 
         mEventId = getIntent().getIntExtra(Constants.EVENT_ID, -1);
         mComments = EventComment.find(EventComment.class, "event_id = ?", String.valueOf(mEventId));
         mCommentAdapter = new CommentAdapter(this, mComments, user, this);
         mRecyclerView.setAdapter(mCommentAdapter);
 
-        findViewById(R.id.ib_send).setOnClickListener(this);
+        mSwipeRefreshLayout = (SwipeRefreshLayout) findViewById(R.id.swipe_refresh_layout);
+        mSwipeRefreshLayout.setColorSchemeResources(R.color.google_blue,
+                R.color.google_green, R.color.google_red, R.color.google_yellow);
+        mSwipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
+            @Override
+            public void onRefresh() {
+                getCommentsTask(mEventId);
+            }
+        });
+        ((NestedScrollView) findViewById(R.id.scrollView))
+                .setOnScrollChangeListener(new NestedScrollView.OnScrollChangeListener() {
+                    @Override
+                    public void onScrollChange(NestedScrollView v,
+                                               int scrollX, int scrollY, int oldScrollX, int oldScrollY) {
+                        if (mNextUrl != null && scrollY == (
+                                v.getChildAt(0).getMeasuredHeight() - v.getMeasuredHeight())) {
+                            mRecyclerView.onScrollStateChanged(RecyclerView.SCROLL_STATE_IDLE);
+                        }
+                    }
+                });
+
+        mRecyclerView.setLoadMoreListener(new LoadMoreListener() {
+            @Override
+            public void onLoadMore() {
+                if (mNextUrl == null) {
+                    mRecyclerView.loadMoreEnd();
+                } else {
+                    loadMoreCommentsTask();
+                }
+            }
+        });
 
         getCommentsTask(mEventId);
+    }
+
+    private void loadMoreCommentsTask() {
+        new EventTaskHandler(this, user.getToken()).loadMoreComments(mNextUrl,
+                new HttpResponseInterface<CommentsResult>() {
+                    @Override
+                    public void onStart() {
+
+                    }
+
+                    @Override
+                    public void onSuccess(CommentsResult classOfT) {
+                        mNextUrl = classOfT.getNext();
+                        for (EventComment comment : classOfT.getComments()) {
+                            saveCommentData(comment);
+                        }
+                        mCommentAdapter.setData(mComments);
+                        mRecyclerView.loadMoreComplete();
+                    }
+
+                    @Override
+                    public void onFailure(HttpBaseResult result) {
+                        mRecyclerView.loadMoreComplete();
+                    }
+
+                    @Override
+                    public void onHttpError(String error) {
+                        mRecyclerView.loadMoreComplete();
+                    }
+                });
     }
 
     private void getCommentsTask(final int eventId) {
@@ -71,16 +181,17 @@ public class CommentActivity extends BaseActivity implements CommentAdapter.Comm
                 new HttpResponseInterface<CommentsResult>() {
             @Override
             public void onStart() {
-
+                mSwipeRefreshLayout.setRefreshing(true);
             }
 
             @Override
             public void onSuccess(CommentsResult classOfT) {
+                mNextUrl = classOfT.getNext();
+                mSwipeRefreshLayout.setRefreshing(false);
                 EventComment.deleteAll(EventComment.class, "event_id = ?", String.valueOf(eventId));
                 mComments = new ArrayList<>();
                 for (EventComment eventComment : classOfT.getComments()) {
-                    eventComment.save();
-                    mComments.add(eventComment);
+                    saveCommentData(eventComment);
                 }
 
                 mCommentAdapter.setData(mComments);
@@ -88,14 +199,21 @@ public class CommentActivity extends BaseActivity implements CommentAdapter.Comm
 
             @Override
             public void onFailure(HttpBaseResult result) {
-
+                mSwipeRefreshLayout.setRefreshing(false);
             }
 
             @Override
             public void onHttpError(String error) {
-
+                mSwipeRefreshLayout.setRefreshing(false);
             }
         });
+    }
+
+    private void saveCommentData(EventComment comment) {
+        EventComment.deleteAll(EventComment.class,
+                "comment_id = ?", String.valueOf(comment.getCommentId()));
+        comment.save();
+        mComments.add(comment);
     }
 
     private void addCommentTask(String text) {
@@ -111,7 +229,6 @@ public class CommentActivity extends BaseActivity implements CommentAdapter.Comm
                     public void onSuccess(EventComment classOfT) {
                         classOfT.save();
                         mCommentAdapter.addData(classOfT);
-                        mRecyclerView.smoothScrollToPosition(mCommentAdapter.getItemCount() - 1);
                         mCommentEditText.setText("");
                         Utils.hideSoftKeyboard(CommentActivity.this);
                     }
@@ -138,9 +255,9 @@ public class CommentActivity extends BaseActivity implements CommentAdapter.Comm
 
                     @Override
                     public void onSuccess(EventComment classOfT) {
-                        mComments.remove(comment);
-                        mCommentAdapter.notifyDataSetChanged();
-                        if (classOfT != null) {
+                        if (classOfT != null && classOfT.getCommentId() == comment.getCommentId()) {
+                            mComments.remove(comment);
+                            mCommentAdapter.notifyDataSetChanged();
                             EventComment.deleteAll(EventComment.class, "comment_id = ?",
                                     String.valueOf(comment.getCommentId()));
                         }
@@ -221,7 +338,7 @@ public class CommentActivity extends BaseActivity implements CommentAdapter.Comm
                             }
                         }
                     })
-                    .setNegativeButton(R.string.cancel, null)
+//                    .setNegativeButton(R.string.cancel, null)
                     .show();
         } else {
             mReplyComment = comment;
